Note for Waterline syntax change
Hi, I added a paragraph in order to note the syntax changes in Waterline queries. I just updated to Waterline 10-rc9 and got a lot of errors until I found out, that the all the final calls on Waterline queries had changed from the keyword .done() to .exec().

At least that's what I found and understood and my code works again, when I make this change, but maybe there's more to it? Would love to get some more info on this change – especially as it seems to be a breaking change.

All the best and thanks for the work you're putting into Sails!
Marcus