Fixes One-to-one association reference
The *one-to-one* association was referencing the behavior of a one-way association. If I understand correctly, as long as you define a *model* association on both models, that makes it *one-to-one* (bidirectional). If you only define the model in one model that makes it **one-to-one** (unidirectional). Maybe I'm misunderstanding something based on [this waterline documenation](https://github.com/balderdashy/waterline-docs/blob/master/associations.md)