Updated onConnect and onDisconnect to conform to latest version
I was not using authorization before so I am not sure how it should be changed. Migration guide said to use beforeConnect if I was using a custom function before, does this mean it would still be called by authorization parameter, when set to true? If so then I suppose this is the correct change, but I can't be sure since I didn't use this functionality.