Document api/hooks/file.js alternative
The Sails code allows a js file in the `api/hooks` directory without a subfolder. Some users (and myself) might prefer to have the `api/hooks` directory full of files instead of have a bunch of folders with one `index.js` file in them. It seems worth mentioning in the documentation.