updated "Tutorial on how to implement passport.js with sails.js" link
The existing tutorial is full of errors. Was unable to get it to work. This "how to" link however is error free and gets the concepts across clearly. Also sails-auth says it is deprecated. 

Also changed order because using passport.js is the simplest way to get auth to work for beginners.