clarify name of "Default" column in table (these are no longer the implicit defaults per se)
In Sails v1, when you do `sails.load()` or `sails.lift()` programmatically in a place outside of a Sails project (i.e. no config files) -- then `sails.config.globals` will be `false`.  But the boilerplate `config/globals.js` has these things set to true.

Worth noting:  there is a special case for sails.config.globals.services -- if sails.config.globals is a dictionary, and the `services` key is undefined, then we default it to `true`.  Otherwise, for the other keys, we default to `false`.  (This is just because we're de-emphasizing services in the Sails 1.0 release)