[misc] Proposing features corrected edit
After reading through "Submitting a proposal" and considering that ROADMAP.md and the Trello board are no longer part of our process, I realized that most of the content on this page would have to be thrown out. The remaining content wasn't going to make a whole lot of sense on its own, so I propose a reorganization in which "Proposing features" is more about the philosophy behind proposals, while "Submitting a proposal" details the actual process (which is much shorter now).