[misc] minor change
Capitalized the first word in the title. 
I would guess that Sails is lowercase in the title because it refers to a command-line command. I've been formatting it in sentence case in my updates, but we could instead use the ticks to format the title as a command and leave everything lowercase. Let me know if that would be better, and I can go back in and reformat the titles within the Command-line interface section.