[misc] changes to res.cookie.md
Ticks added to title
In Details, removed the reference to "the previous example", which doesn't seem to exist. 
It seems like a lot of the terms in double quotes should be in ticks, based on how I've seen them (double quotes and ticks) used in the docs thus far.
I think that bodyParser() middleware is referring to the Express body-parsing middleware, so I made that explicit. 
Reworded a couple of things for clarity and flow.